\name{googleanalytics4r-package}
\Rdversion{1.1}
\alias{googleanalytics4r-package}
\alias{googleanalytics4r}
\docType{package}
\title{
  Google Analytics API for R
}
\description{
  Functions to acccess the Google Analytics API in R.
}
\details{
  \tabular{ll}{
    Package: \tab googleanalytics4r\cr
    Type: \tab Package\cr
    Version: \tab 0.1-0\cr
    Date: \tab 2009-11-17\cr
    License: \tab GPL Version 3 or later\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Soren Macbeth

  Maintainer: Soren Macbeth <soren@dopeness.org>
}
\references{
  http://code.google.com/apis/analytics/docs/
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
